Fix pdfmaker support for colors specified in print config
Use the actual calculated fill color when generating a textbox instead
of hardcoding `'black'`.

The changes in regards to the pdf titlepage did break the custom colors
specified in the print configuration when generating a pdf. Notably, the
synopsis was always rendered in black instead of the ~50% grey. But this
also affects the header and footer that are rendered with explicit color
values.

Related: b30bf507e492285e7325d72a800b0e1eba3af0e0